migrate settings from profile to settings dapp
- [CORE-1129]
