+++ dapps/profile.vue/src/i18n/de.json
@@ -10,7 +10,6 @@
       "detail": "Identität",
       "did": "Decentralized ID",
       "logout": "Ausloggen",
-      "settings": "Einstellungen",
       "sharings": "Freigaben",
       "verifications": "Verifizierungen",
       "wallet": "Wallet"
@@ -195,32 +194,6 @@
     },
     "not-permitted-view": "Sie haben keine Rechte, diese Daten für die ausgewählte Identität anzusehen.",
     "profile": "Profil",
-    "security-info": {
-      "encryption-key": {
-        "exported": "Encryption-Key in die Zwischenablage kopiert",
-        "title": "Encryption Key kopieren"
-      },
-      "private-key": {
-        "exported": "Private-Key in die Zwischenablage kopiert",
-        "title": "Private Key kopieren"
-      },
-      "title": "Security Information"
-    },
-    "settings": {
-      "desc": "Einstellungen",
-      "dev-domain": {
-        "desc": "Mit dieser Konfiguration wird die Endung \".evan\" durch die angegebene Adresse ersetzt. Über diese Logik ist es möglich, Applikation auf verschiedenen Domains gleichzeitig für Testzwecke zu verwenden (z.B.: myapp.test.evan).",
-        "title": "Applikation von Test-Domain"
-      },
-      "developer-mode": "Entwicklermodus",
-      "language": "Sprache",
-      "languages": {
-        "browser": "Browser-Sprache",
-        "de": "Deutsch",
-        "en": "Englisch"
-      },
-      "reload-hint": "Bitte laden Sie die Anwendung neu, wenn Sie die Spracheinstellungen geändert haben."
-    },
     "sharing": {
       "description": "Um Ihre Identitätsdaten mit anderen Partnern zu teilen, wählen Sie zunächst den entsprechenden Kontakt sowie danach die zu teilenden Daten aus. Klicken Sie auf \"Daten freigeben\".",
       "description-no-contacts": "Identitätsinformationen können mit anderen Partnern geteilt werden, nachdem Sie eine sichere Verbindung zu deren evan.network Identität aufgebaut haben. Gehen Sie dazu in die Kontaktübersicht und laden Sie Ihre Partner per E-Mail oder auch per Identitäts-ID ein, um Ihre Daten vertrauensvoll zu teilen.",
