+++ dapps/settings/src/components/account/account.ts
@@ -22,7 +22,47 @@ import Component, { mixins } from 'vue-class-component';
 
 // evan.network imports
 import { EvanComponent } from '@evan.network/ui-vue-core';
+import { session, lightwallet } from '@evan.network/ui-session';
 
 @Component({ })
 export default class AccountSettingsComponent extends mixins(EvanComponent) {
+  // TODO refactor to (renderless) vue component
+  copyToClipboard(text: string, toastText: string): void {
+    const textArea = document.createElement('textarea');
+
+    textArea.value = text;
+    document.body.appendChild(textArea);
+    textArea.focus();
+    textArea.select();
+    document.execCommand('copy');
+    document.body.removeChild(textArea);
+
+    this.$toasted.show(
+      this.$t(toastText),
+      {
+        duration: 3000,
+        type: 'success',
+      },
+    );
+  }
+
+  /**
+   * Export the private key for the current logged in account
+   */
+  async exportPrivateKey(): Promise<void> {
+    const vault = await lightwallet.loadUnlockedVault();
+    const privateKey = lightwallet.getPrivateKey(vault, session.activeAccount);
+
+    this.copyToClipboard(privateKey, '_profile.security-info.private-key.exported');
+  }
+
+  /**
+   * Export the private key for the current logged in account
+   */
+  async exportEncryptionKey(): Promise<void> {
+    await lightwallet.loadUnlockedVault();
+    const encryptionKey = await lightwallet.getEncryptionKey();
+
+    this.copyToClipboard(encryptionKey, '_profile.security-info.encryption-key.exported');
+  }
 }
