+++ dapps/settings/src/components/browser/browser.vue
@@ -18,13 +18,80 @@
 */
 
 <template>
-  <div>
-    browser config
+  <div class="evan-content-container">
+    <div class="white-box mt-5">
+      <h3 class="mb-5 d-block">
+        {{ `_settings.browser.language` | translate }}
+      </h3>
+      <select
+        id="language"
+        ref="language"
+        v-model="language"
+        class="form-control custom-select"
+        @change="languageChanged()"
+      >
+        <option value="">
+          {{ '_settings.browser.languages.browser' | translate }}
+        </option>
+        <option value="de">
+          {{ '_settings.browser.languages.de' | translate }}
+        </option>
+        <option value="en">
+          {{ '_settings.browser.languages.en' | translate }}
+        </option>
+      </select>
+      <span class="text-muted small">
+        {{ '_settings.browser.reload-hint' | translate }}
+      </span>
+    </div>
+
+    <div class="white-box mt-5">
+      <h3 class="mb-5 d-block">
+        {{ '_settings.browser.developer-mode' | translate }}
+      </h3>
+      <evan-form-control-checkbox
+        id="devMode"
+        v-model="devMode"
+        class="mt-3"
+        style="min-width: 0;"
+        @input="devModeChanged()"
+      />
+
+      <template v-if="devMode">
+        <b>
+          {{ `_settings.browser.dev-domain.title` | translate }}
+        </b>
+        <div class="d-flex align-items-center mt-3">
+          <div>
+            <evan-form-control-checkbox
+              id="devDomainEnabled"
+              v-model="devDomainEnabled"
+              class="mr-3"
+              @input="devDomainChanged()"
+            />
+          </div>
+          <div class="form-group ml-3 w-100">
+            <input
+              id="devDomain"
+              ref="devDomain"
+              v-model="devDomain"
+              class="form-control"
+              :placeholder="`_settings.browser.dev-domain.desc` | translate"
+              :disabled="!devDomainEnabled"
+              @change="devDomainChanged()"
+            >
+          </div>
+        </div>
+        <span class="text-muted small">
+          {{ '_settings.browser.dev-domain.desc' | translate }}
+        </span>
+      </template>
+    </div>
   </div>
 </template>
 
 <script lang="ts">
-import Component from './browser.ts';
+import Component from './browser';
 
 export default Component;
 </script>
